feat: implement production reporting repository and calving projection use case

+++ src/main/java/com/ganaderia/application/dto/ReporteProduccionDTO.java
@@ -1,13 +1,11 @@
 package com.ganaderia.application.dto;
 
-import java.math.BigDecimal;
-
 public class ReporteProduccionDTO {
     private String vacaId;
     private String numeroArete;
-    private BigDecimal totalLitros;
+    private Double totalLitros;
 
-    public ReporteProduccionDTO(String vacaId, String numeroArete, BigDecimal totalLitros) {
+    public ReporteProduccionDTO(String vacaId, String numeroArete, Double totalLitros) {
         this.vacaId = vacaId;
         this.numeroArete = numeroArete;
         this.totalLitros = totalLitros;
@@ -21,7 +19,7 @@ public class ReporteProduccionDTO {
         return numeroArete;
     }
 
-    public BigDecimal getTotalLitros() {
+    public Double getTotalLitros() {
         return totalLitros;
     }
 }
